backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
openai==1.51.0
python-dotenv==1.0.1
